pyproject.toml
[project]
name = "twincli"
version = "0.1.3"
description = "A lightweight Gemini terminal client with a twin-pun name"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-generativeai>=0.4.1",
    "rich",
    "click",
    "selenium>=4.0.0",
    "requests"
]

[project.scripts]
twincli = "twincli.__main__:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"